requirements_dev.txt
iniconfig==2.0.0
numpy==1.26.0
packaging==23.1
pandas==2.1.1
pluggy==1.3.0
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
tzdata==2023.3
